fix: adding interface and impl for WishlistService
