+++ src/main/java/co/simplon/wishmegift/service/WishlistService.java
@@ -1,35 +1,15 @@
 package co.simplon.wishmegift.service;
 
 import co.simplon.wishmegift.entity.Wishlist;
-import co.simplon.wishmegift.repository.WishlistRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class WishlistService {
+public interface WishlistService {
+    Wishlist create(Wishlist wishlist);
 
-    private final WishlistRepository wishlistRepository;
+    List<Wishlist> getAll();
 
-    public WishlistService(WishlistRepository wishlistRepository) {
-        this.wishlistRepository = wishlistRepository;
+    Wishlist getById(Long id);
 
-    }
-
-    public Wishlist create(Wishlist wishlist) {
-        return wishlistRepository.save(wishlist);
-    }
-
-    public List<Wishlist> getAll() {
-        return wishlistRepository.findAll();
-    }
-
-    public Wishlist getById(Long id) {
-        return wishlistRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Liste introuvable"));
-    }
-
-    public void delete(Long id) {
-        wishlistRepository.deleteById(id);
-    }
+    void delete(Long id);
 }
